Use rev option to set fixed commit hash
This change sets a fixed commit revision to the `riscv` patch annotation
in all `Cargo.toml` files.

Using `master` as a `branch` option results in different commit hashes at different times, depending on changes in the `riscv` crate. This crate has seen some changes in the last days / weeks that broke the current setup, therefore a working revision is set.